+++ src/main/java/org/realityforge/replicant/server/transport/ReplicantSession.java
@@ -1,5 +1,9 @@
 package org.realityforge.replicant.server.transport;
 
+import javax.annotation.Nonnull;
+import javax.json.stream.JsonGenerator;
+import org.realityforge.rest.field_filter.FieldFilter;
+
 /**
  * Base class for sessions within replicant.
  */
@@ -17,4 +21,13 @@ public abstract class ReplicantSession
   {
     return _queue;
   }
+
+  /**
+   * Emit replication status of the session.
+   * This should include details of subscriptions.
+   *
+   * @param g the json generator in which to serialize details.
+   * @param filter the filter to use when determining which details to emit.
+   */
+  public abstract void emitStatus( @Nonnull final JsonGenerator g, @Nonnull FieldFilter filter );
 }